BUG: Arreglar bug no se muestran viajes

+++ app/src/main/java/eina/unizar/frontend/models/Viaje.kt
@@ -11,7 +11,7 @@ data class Viaje(
     val fechaHoraInicio: String, // ISO 8601, se puede parsear a LocalDateTime si queremos
     val fechaHoraFin: String,
     val kmRealizados: Int,
-    val consumoCombustible: Int,
+    val consumoCombustible: Double,
     val ubicacionFinal: Ubicacion,
     val usuario: String
 )
